Code Refactor: Se centraliza la logica de autenticacion
• Se modifica el .gitignore
• Se centraliza la logica de autenticacion dentro del archivo auth.py del directorio routers
